bug fix & style change
See the previous two commits for a segmented overview of what the (hidden - not obvious from the tests) bug was. I noticed this bug while thinking through how to do Radix MSB. The effect of the bug, if it wasn't caught, would have been items missing in the output from larger input sizes than what are shown in the test script. Note to self to add some large inputs to the radix test script in the future. In this commit, I have set j < dVct which is functionally the same as j < nRows, but I think that it is more readable. I have also made the memset calls more readable.